LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/at-container-parsing-expected.txt


PASS size(width)
PASS size(min-width: 0px)
PASS size(max-width: 0px)
PASS size(height)
PASS size(min-height: 0px)
PASS size(max-height: 0px)
PASS size(aspect-ratio)
PASS size(min-aspect-ratio: 1/2)
PASS size(max-aspect-ratio: 1/2)
PASS size(orientation: portrait)
PASS size(inline-size)
PASS size(min-inline-size: 0px)
PASS size(max-inline-size: 0px)
PASS size(block-size)
PASS size(min-block-size: 0px)
PASS size(max-block-size: 0px)
PASS size(width: 100px)
PASS size((width: 100px))
PASS size(not (width: 100px))
PASS size((width: 100px) and (height: 100px))
PASS size((width: 50px) or (height: 100px))
PASS size(width < 100px)
PASS size(100px < width)
PASS size(100px < width < 200px)
FAIL foo(width) assert_equals: expected "" but got "true"
FAIL size(asdf) assert_equals: expected "" but got "true"
FAIL size(resolution > 100dpi) assert_equals: expected "" but got "true"
FAIL size(resolution: 150dpi) assert_equals: expected "" but got "true"
FAIL size(color) assert_equals: expected "" but got "true"
FAIL size(min-color: 1) assert_equals: expected "" but got "true"
FAIL size(color-index >= 1) assert_equals: expected "" but got "true"
FAIL (color-index >= 1) assert_equals: expected "" but got "true"
FAIL size(grid) assert_equals: expected "" but got "true"
FAIL (grid) assert_equals: expected "" but got "true"
PASS screen
PASS print
PASS not print
PASS only print
PASS screen and (width: 100px)
PASS screen or (width: 100px)
PASS not screen and (width: 100px)
PASS not screen or (width: 100px)
FAIL (width: 100px), (height: 100px) assert_equals: expected 0 but got 1
PASS Container selector:  foo
PASS Container selector:  foo
PASS Container selector: name(foo)
PASS Container selector: name( foo )
PASS Container selector: type(size)
PASS Container selector: type( size )
PASS Container selector: type(inline-size)
PASS Container selector: type(block-size)
PASS Container selector: name(bar) type(block-size)
PASS Container selector: type(block-size) name(bar)
PASS Container selector: foo foo
PASS Container selector: 1px
PASS Container selector: 50gil
FAIL Container selector: name(1px) assert_equals: expected 0 but got 1
FAIL Container selector: type(1px) assert_equals: expected 0 but got 1
FAIL Container selector: type(red) assert_equals: expected 0 but got 1
FAIL Container selector: type(size) type(size) assert_equals: expected 0 but got 1
FAIL Container selector: type(inline-size) type(block-size) assert_equals: expected 0 but got 1
FAIL Container selector: name(foo) name(bar) assert_equals: expected 0 but got 1
FAIL Container selector: type(inline-size) name(foo) type(block-size) assert_equals: expected 0 but got 1
FAIL Container selector: name(foo) type(size) name(bar) assert_equals: expected 0 but got 1
FAIL Container selector: name(type(size)) assert_equals: expected 0 but got 1
FAIL Container selector: type(name(bar)) assert_equals: expected 0 but got 1
